--- 6-AdvancedScenarios/4-hybrid-SPA/App/msal.js
@@ -1,21 +1,0 @@
-const msal = require('@azure/msal-node');
-require('dotenv').config();
-
-const msalInstance = new msal.ConfidentialClientApplication({
-    auth: {
-        clientId: process.env.CLIENT_ID,
-        authority: `https://login.microsoftonline.com/${process.env.TENANT_ID}`,
-        clientSecret: process.env.CLIENT_SECRET
-    },
-    system: {
-        loggerOptions: {
-            loggerCallback: (loglevel, message, containsPii) => {
-                console.log(message);
-            },
-            piiLoggingEnabled: false,
-            logLevel: msal.LogLevel.Verbose,
-        }
-    }
-});
-
-module.exports = msalInstance;
