+++ 6-AdvancedScenarios/4-sign-in-hybrid/App/client/public/manifest.json
