+++ 6-AdvancedScenarios/4-sign-in-hybrid/App/client/src/App.test.js
@@ -3,6 +3,6 @@ import App from './App';
 
 test('renders learn react link', () => {
   render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
+  const linkElement = screen.getByText(/Microsoft identity platform/i);
   expect(linkElement).toBeInTheDocument();
 });
