+++ 6-AdvancedScenarios/4-sign-in-hybrid/App/client/src/index.js
@@ -2,7 +2,6 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { App } from "./App.jsx";
 
-import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
 import { PublicClientApplication } from "@azure/msal-browser";
 import { msalConfig } from "./authConfig";
@@ -18,8 +17,3 @@ ReactDOM.render(
   </BrowserRouter>,
   document.getElementById("root")
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
