+++ 6-AdvancedScenarios/4-sign-in-hybrid/App/client/src/styles/index.css
