+++ 6-AdvancedScenarios/4-sign-in-hybrid/App/package.json
