+++ 6-AdvancedScenarios/4-sign-in-hybrid/AppCreationScripts/Cleanup.ps1
