+++ 6-AdvancedScenarios/4-sign-in-hybrid/AppCreationScripts/Configure.ps1
